Move args parsing outside src. Add tests main
