+++ main_tests.py
@@ -1,0 +1,7 @@
+from src.tests import sound_from_file_test
+from src.tests import tonation_from_sounds_test
+
+
+if __name__ == "__main__":
+    sound_from_file_test.run_tests()
+    tonation_from_sounds_test.run_tests()
