+++ src/__init__.py
@@ -1,1 +1,1 @@
-# flake8: noqa 
+# flake8: noqa
